integrate a first version without any error handling
