+++ milli/Cargo.toml
@@ -39,8 +39,8 @@ tempfile = "3.2.0"
 uuid = { version = "0.8.2", features = ["v4"] }
 
 # facet filter parser
+filter_parser = { path = "../filter_parser" }
 nom = "7.0.0"
-nom_locate = "4.0.0"
 
 # documents words self-join
 itertools = "0.10.0"
