+++ milli/src/lib.rs
@@ -34,7 +34,7 @@ pub use self::heed_codec::{
     RoaringBitmapLenCodec, StrBEU32Codec, StrStrU8Codec,
 };
 pub use self::index::Index;
-pub use self::search::{FacetDistribution, FilterCondition, MatchingWords, Search, SearchResult};
+pub use self::search::{FacetDistribution, Filter, MatchingWords, Search, SearchResult};
 
 pub type Result<T> = std::result::Result<T, error::Error>;
