+++ milli/src/search/facet/mod.rs
@@ -1,10 +1,9 @@
 pub use self::facet_distribution::FacetDistribution;
 pub use self::facet_number::{FacetNumberIter, FacetNumberRange, FacetNumberRevRange};
 pub use self::facet_string::FacetStringIter;
-pub use self::filter_condition::FilterCondition;
+pub use self::filter_condition::Filter;
 
 mod facet_distribution;
 mod facet_number;
 mod facet_string;
 mod filter_condition;
-mod filter_parser;
